pyproject.toml
[project]
name = "palantir"
version = "0.1.0"
description = "AI-powered classroom assistant for Raspberry Pi 5"
requires-python = ">=3.11"
dependencies = [
    # Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0",
    # Audio
    "sounddevice>=0.5.0",
    "numpy>=1.26.0",
    # Vision
    "opencv-python-headless>=4.10.0",
    # Data
    "redis[hiredis]>=5.2.0",
    "pydantic>=2.10.0",
    # Config
    "python-dotenv>=1.0.0",
    # Hardware
    "gpiozero>=2.0.0",
    # Utilities
    "structlog>=24.0.0",
    # TLS certificate generation (self-signed for local-network HTTPS)
    "cryptography>=43.0.0",
]

[project.optional-dependencies]
# Phase 2: Voice pipeline
voice = [
    "anthropic>=0.42.0",
    "faster-whisper>=1.1.0",
    "openwakeword>=0.6.0",
    "piper-tts>=1.2.0",
    "onnxruntime>=1.19.0",
]
# Phase 3: Facial recognition
face = [
    "insightface>=0.7.0",
    "onnxruntime>=1.19.0",
]
# Phase 4: Speaker ID
speaker = [
    "speechbrain>=1.0.0",
]
# Phase 5: Object detection
objects = [
    "ultralytics>=8.3.0",
]
# All ML dependencies
ml = [
    "palantir[voice,face,speaker,objects]",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    # In-process Redis shim for testing without a real redis-server
    "fakeredis>=2.23.0",
]

[project.scripts]
palantir-audio = "palantir.audio.service:main"
palantir-vision = "palantir.vision.service:main"
palantir-brain = "palantir.brain.service:main"
palantir-tts = "palantir.tts.service:main"
palantir-eventlog = "palantir.eventlog.service:main"
palantir-web = "palantir.web.main:main"
palantir-backup = "palantir.backup:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/palantir"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
